==> 10_Extracting covariates.R <==
library(raster)
library(rstudioapi)
library(tidyverse)
library(sf)
library(dplyr)
library(sp)
library(rgdal)
library(rgeos)
library(spatialEco)

##Set working directory----
## Set work directory----
working.dir <- dirname(rstudioapi::getActiveDocumentContext()$path) # sets working directory to where this script is saved (DON't MOVE THE SCRIPT)

## Set sub directories----
d.dir <- paste(working.dir,"Tidy data",sep="/") 
s.dir <- paste(working.dir,"Spatial",sep="/") # spatial is where I keep spatial data files, rasters and shapefiles
p.dir <- paste(working.dir,"Plots",sep="/")

lat.long <- "+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"
utm <- "+proj=utm +zone=49 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"


## Load bathymetry and tpi data----
setwd(s.dir)
bathy <- raster(paste(s.dir, "bathy.tif", sep='/'))
bathy <- flip(bathy, direction="y")
plot(bathy) # check it
proj4string(bathy) # check the coordinate system, resolution, etc..


tpi <- raster(paste(s.dir, "tpi.tif", sep='/'))
plot(tpi) # check it
proj4string(tpi) # check the coordinate system, resolution, etc..
tpi <- flip(tpi, direction="y")
plot(tpi)

## Load BRUV data
bruv <-  read.csv(paste(d.dir, "ningaloo_metadata.csv", sep ='/'))
str(bruv) # check df
names(bruv) # check the names of the columns with lat and lon
# Extract bits we need
bruv <- bruv %>%
  select("sample", "longitude", "latitude", "site")
bruv


# Turn bruv data into spatial points --
coordinates(bruv) <- ~ longitude + latitude
points(bruv) # check it - this should plot the points on top of your previously plotted tpi or bathy
proj4string(bruv) # check coordinate system and/or projection
# If no projection give assign one,
proj4string(bruv) <- lat.long
# Convert to the same projection as the bathy/tpi
# Rasters and spatial points should have the sample coordinate system and/or projection
bruv <- spTransform(bruv, CRS("+proj=utm +zone=49 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"))
points(bruv)

## Create all covariates wanted - already have bathymetry and TPI

Slope <- raster::terrain(bathy, opt="slope")
writeRaster(Slope, paste(s.dir, "slope.tif", sep ='/'))
Aspect <- raster::terrain(bathy, opt="aspect")
writeRaster(Aspect, paste(s.dir, "aspect.tif", sep ='/'))
TRI <- raster::terrain(bathy, opt="tri")
# writeRaster(preds, paste(s.dir, "TRI.tif", sep ='/'))
Roughness <- raster::terrain(bathy, opt="roughness")
writeRaster(Roughness, paste(s.dir, "roughness.tif", sep ='/'))
FlowDir <- raster::terrain(bathy,opt="flowdir")
writeRaster(FlowDir, paste(s.dir, "flowdir.tif", sep ='/'))

## Extract predictor data at bruv locations ----

# Stack predictors --
preds <- raster::stack(bathy,tpi,Slope,Aspect,TRI,Roughness,FlowDir) # if this fails: preds may have different extent or resolution

# Extract predictors info --
covariates <- raster::extract(preds, bruv, df = T) # this will result in dataframe with bruv points and predictor values

# Give bruv uniques IDs
bruv$ID <- covariates$ID

# Fix column names
names(covariates) <- c("ID", "bathymetry", "TPI", "Slope", "Aspect","TRI", "Roughness", "FlowDir")

# Create full dataframe
fullcovariates <- merge.data.frame(covariates, bruv, by="ID")

# Save this data frame
write.csv(fullcovariates, paste(d.dir, "covariates.csv", sep ='/')) 

# Save stack of predictors if desired
writeRaster(preds, paste(s.dir, "covariatestack.tif", sep ='/'))

###### Finding NA values #######

NA.bathy <- subset(fullcovariates,is.na(fullcovariates$bathymetry))
NA.tpi <- subset(fullcovariates,is.na(fullcovariates$TPI))
NA.slope <- subset(fullcovariates,is.na(fullcovariates$Slope))
NA.aspect <- subset(fullcovariates,is.na(fullcovariates$Aspect))
NA.tri <- subset(fullcovariates,is.na(fullcovariates$TRI))
NA.roughness <- subset(fullcovariates,is.na(fullcovariates$Roughness))
NA.flowdir <- subset(fullcovariates,is.na(fullcovariates$FlowDir))

unique(NA.bathy$sample) #10.12
unique(NA.tpi$sample) # 8.05 10.09 10.12 16.03
unique(NA.slope$sample) #10.12 16.03
unique(NA.aspect$sample) #10.12 16.03
unique(NA.tri$sample) #10.12 16.03
unique(NA.roughness$sample) #10.12 16.03
unique(NA.flowdir$sample) #10.12

## Plots of NA sites 

#Bathymetry
plot(bathy)
points(bruv, pch=20, cex=0.75, col=ifelse(bruv$sample=='10.12', "red", "black"))

#TPI
plot(tpi)
points(bruv, pch=20, cex=0.75, col=ifelse(bruv$sample %in% c("8.05", "10.09", "10.12", "16.03"), "red", "black"))

#Slope
plot(Slope)
points(bruv, pch=20, cex-0.75, col=ifelse(bruv$sample %in% c("10.12", "16.03"), "red", "black"))

#Aspect
plot(Aspect)
points(bruv, pch=20, cex-0.75, col=ifelse(bruv$sample %in% c("10.12", "16.03"), "red", "black"))

#TRI
plot(TRI)
points(bruv, pch=20, cex-0.75, col=ifelse(bruv$sample %in% c("10.12", "16.03"), "red", "black"))

#Roughness
plot(Roughness)
points(bruv, pch=20, cex-0.75, col=ifelse(bruv$sample %in% c("10.12", "16.03"), "red", "black"))

#FlowDir
plot(FlowDir)
points(bruv, pch=20, cex=0.75, col=ifelse(bruv$sample=='10.12', "red", "black"))

############# Formating habitat data to use as covariate #############

setwd(d.dir)

habitat <- read.csv("ningaloo.complete.habitat.csv")

## Want corals and sponges to stay as coral/sponge when coverage is greater than 50% otherwise we want it to
## be classified as just reef but if unconsolidated sediment is >50% we want it to be classified as unconsolidated

names(habitat)
habitat.classified <- habitat %>%
  rowwise()%>%
  mutate(reef=sum(c(broad.bryozoa,broad.crinoids,broad.hydrocoral,broad.hydroids,broad.octocoral.black,broad.sponges)))%>%
  mutate(sand=broad.unconsolidated)

habitat.classified <- habitat.classified%>%
  select("sample", "mean.relief", "sd.relief", "reef", "sand")

write.csv(habitat.classified, "habitat.classified.csv")

# Add to existing covariates
covariates<-read.csv("covariates.csv")

covariates<-full_join(covariates,habitat.classified,by="sample")

############# Adding distance to boat ramp to covariates ##############
ramps<-read.csv("distance.to.ramp.csv")

covariates<-full_join(covariates,ramps,by="sample")
  
covariates<-covariates%>%
  select(!c("X.x", "X.y"))%>%
  dplyr::rename(distance.to.ramp=Distance.to.ramp)

covariates<-covariates[,c(1,9,10,11,12,2,3,4,5,6,7,8,13,14,15,16,17)]

write.csv(covariates, "covariates.csv")

############## Adding status of site to covariates ##############
setwd(d.dir)
covariates <- read.csv('covariates.csv')


covariates <- covariates%>%
  mutate(status= 
           if_else(sample%in%c('5.02','5.03','5.07','5.08','5.11','5.13','6.01','6.02','6.03','6.04','6.05',
                              '6.12','7.01','7.02','7.03','7.04','7.05','7.08','8.01','8.02','8.05','8.06',
                              '8.07','8.12','9.01','9.02','9.05','9.06','9.07','9.08','10.01','10.02','10.03',
                              '10.04','10.07','10.08', '10.12', '10.13', '10.14', '10.15', '14.02', '14.12', 
                              '14.13'),"NT","F"))




covariates <- covariates%>%
  mutate(zone=
           if_else(sample%in%c("1.01", "1.02", "1.03", "1.04", "1.05", "1.06", "3.01", "3.02", "3.03", "3.04",
                               "3.05", "3.06", "4.01", "4.02", "4.03", "4.08", "4.11", "4.12"), "SC", ifelse(
                                 sample%in%c('5.02','5.03','5.07','5.08','5.11','5.13','6.01','6.02','6.03','6.04','6.05',
                                             '6.12','7.01','7.02','7.03','7.04','7.05','7.08','8.01','8.02','8.05','8.06',
                                             '8.07','8.12','9.01','9.02','9.05','9.06','9.07','9.08','10.01','10.02','10.03',
                                             '10.04','10.07','10.08'), "NP", ifelse(sample%in%c("11.01", "11.02", "11.03",
                                                                                                "11.04", "11.05", "11.06",
                                                                                                "12.02", "12.03", "12.04",
                                                                                                "12.04", "12.05", "12.05",
                                                                                                "12.11", "13.02", "13.03",
                                                                                                "13.04", "13.05", "13.06",
                                                                                                "13.08", "14.11"), "NC", "SA")
                               )))

write.csv(covariates, "covariates.csv")
